api/requirements.txt
numpy==1.22.2
uvicorn==0.17.5
Pillow==9.0.1
tensorflow==2.7.0
fastapi==0.74.0
python-multipart==0.0.5